requirements.txt
tk
pillow
PyMuPdf
fitz